Add useEditPost hook
